pyproject.toml
[tool.poetry]
name = "transfold-bot"
version = "0.1.0"
description = ""
authors = ["amirzenoozi <amirzenoozi72@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
python-dotenv = "^1.2.2"
python-telegram-bot = "^22.6"
asyncio = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
